+++ lib/pixel_color_image.dart
@@ -1,3 +1,4 @@
 export 'package:pixel_color_image/src/pixel_color_image.dart';
 export 'package:pixel_color_image/src/get_image_bytes_url.dart';
 export 'package:pixel_color_image/src/get_image_bytes_asset.dart';
+export 'package:pixel_color_image/src/pixel_color.dart';