+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 ## 1.0.0
 
 - First Release
+
+## 1.0.1
+
+- fix documents
+
+## 1.1.0
+
+- add: PixelColor.assetImage(...)
+- add: PixelColor.networkImage(...)
+
+## 1.1.1
+
+- fix PixelColor